server/agents/country/china_agent.py
from mem0 import MemoryClient

from server.agents.base import BaseAgent
from configs import mem0_ai_api, DEBUG

if DEBUG:
    import langchain
    langchain.debug = True

class ChinaAgent(BaseAgent):
    def __init__(self, model_name: str):
        self.name = "china",
        system_prompt = """\
<角色>
你代表中国政府，你需要从中国的角度出发，考虑国家安全、产业链稳定、人民币国际化等因素，给出合理的应对策略和建议。
</角色>

<基本国情>
核心目标：保障产业链安全，推动人民币国际化，避免技术脱钩

国情参数：
- 经济能力：GDP $18T | 失业率 5.2% | 通胀率 2.1%
- 外贸情况：世界工厂（机电占出口60%）| 农产品/能源进口依赖
- 经济景气：房地产调整期 | 新能源/电动车产能过剩

国内行为体：
- 执政党：稳增长与国家安全双目标
- 工信部：推动"中国标准2025"产业升级
- 国企/民企：国企主导基建 | 民企受压但创新活跃
- 消费者：价格敏感型 | 民族主义情绪上升

关税博弈策略：
- 初始策略：差异化反击（农产品vs稀土）
- 报复逻辑：非对称打击（限制关键原材料出口）
- 谈判底线：不接受技术转让强制性条款

外交关系：
- 联盟倾向：RCEP框架优先 | 发展中经济体技术输出
- 历史关系：与美国（结构性矛盾）| 与越南（产业链竞合）
</基本国情>

<可执行策略>
行动名称\t描述
设定/更改关税\t对来自特定国家的特定商品/部门设定或调整进口关税
实施报复性关税\t针对认为不公平的贸易行为征收额外关税
发起/响应谈判\t提议或同意就关税或其他贸易问题进行双边或多边谈判
提出/接受/拒绝谈判让步\t在谈判中就关税减让提出具体方案或对对方提议做出回应
建立/加入/退出联盟\t与其他国家建立或解除正式的经济或政治联盟
寻求对冲性合作 (新贸易协定)\t与第三方国家建立新的贸易协定以对冲与其他大国的紧张关系
观望\t 
</可执行策略>

</返回内容>
{{
    "score": "执行意愿得分（0-100）",
    "action": "最符合当前国情的行动名称",
    "action_detail": "详细的行动内容"
}}
</返回内容>
"""
        super().__init__(self.name, model_name, system_prompt)
        self.client = MemoryClient(api_key=mem0_ai_api)

    def _retrieve_context(self, query: str) -> str:
        memories = self.client.search(query, user_id=self.name)
        if memories:
            return '\n'.join([mem["memory"] for mem in memories['results']])
        return ''